requirements.txt
anyio==4.2.0
attrs==22.1.0
azure-common==1.1.28
azure-core==1.29.6
azure-identity==1.15.0
azure-mgmt-compute==30.4.0
azure-mgmt-consumption==10.0.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
beautifulsoup4==4.12.2
boto3==1.34.14
cachetools==5.3.2
certifi==2022.12.7
cffi==1.14.6
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
cryptography==3.4.8
docutils==0.15.2
ec2instanceconnectcli==1.0.2
ecs-logging==2.0.0
elastic-transport==8.4.0
elasticsearch==7.17.7
elasticsearch8==8.6.2
exceptiongroup==1.2.0
future==0.18.2
google==3.0.0
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-billing==1.12.0
google-cloud-resource-manager==1.11.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.4
isodate==0.6.1
jmespath==0.10.0
jinja2==3.1.3
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
oauth==1.0.1
oauthlib==3.2.2
packaging==23.2
portalocker==2.8.2
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycparser==2.20
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.1
python-dotenv==1.0.0
pytz==2021.1
requests==2.28.1
requests-aws4auth==1.1.2
requests-oauthlib==1.3.1
rsa==4.7.2
six==1.16.0
slack-sdk==3.26.1
sniffio==1.3.0
soupsieve==2.5
tabulate==0.8.10
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.4
voluptuous==0.13.1
